# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's MrSoft News",
        "description": "API's for MrSoft News",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://mrsoft-news.test",
            "description": "Servidor de la API (dinámico: usa APP_URL del entorno actual)"
        }
    ],
    "paths": {
...
        "/api/comment": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get all comments",
                "operationId": "c561bf3062eb2fd096a331aaca319c76",
                "parameters": [
                    {
                        "name": "product",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "Gesrest",
                                "360sys",
                                "HotelHUB",
                                "Comprobante-e",
                                "Mr. Soft"
                            ]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit of clients",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResourceCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Comment"
                ],
                "summary": "Create a new comment/testimonial",
                "operationId": "8aa076dd8e387ea517f9722d89e09af3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "person"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Great product!"
                                    },
                                    "person": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "CEO"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comment/{id}": {
            "get": {
                "tags": [
                    "Comment"
                ],
                "summary": "Get comment by id",
                "operationId": "b14d2a3c17aa747dc601c54bdf348bf9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Comment"
                ],
                "summary": "Update a comment/testimonial",
                "operationId": "58d5268605f3286ab81e3cf9f422a833",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "example": "Great product!"
                                    },
                                    "person": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "CEO"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Comment"
                ],
                "summary": "Delete a comment/testimonial",
                "operationId": "3b25b6f44049d90c415f7b62ca01b468",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Comment not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "AddressResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Address name"
                    }
                },
                "type": "object"
            },
            "AddressResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/AddressResource"
                }
            },
            "ClientResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "News title"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "News description"
                    },
                    "logo": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/storage/logo.png"
                    },
                    "departamento": {
                        "type": "string",
                        "example": "News description"
                    },
                    "type": {
                        "type": "string",
                        "example": "News description"
                    },
                    "departments": {
                        "$ref": "#/components/schemas/DepartmentResourceCollection"
                    },
                    "addresses": {
                        "$ref": "#/components/schemas/AddressResourceCollection"
                    }
                },
                "type": "object"
            },
...
            "CommentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "content": {
                        "type": "string",
                        "example": "Muy buena plataforma me facilita mucho para llevar..."
                    },
                    "person": {
                        "type": "string",
                        "example": "Jose Santos More Monja"
                    },
                    "position": {
                        "type": "string",
                        "example": "Administrador"
                    },
                    "producto": {
                        "type": "string",
                        "example": "Mr. Soft"
                    },
                    "cliente": {
                        "$ref": "#/components/schemas/ClientResource"
                    }
                },
                "type": "object"
            },
            "CommentResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/CommentResource"
                }
            },
            "DepartmentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Department name"
                    }
                },
                "type": "object"
            },
            "DepartmentResourceCollection": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/DepartmentResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}